+++ LeagueOptimizer.Abstractions/Champions/Data/StatData.cs
@@ -2,6 +2,6 @@ namespace LeagueOptimizer.Abstractions.Champions.Data;
 
 public class StatData
 {
-    public double Base { get; set; }
-    public double Growth { get; set; }
+    public decimal Base { get; set; }
+    public decimal Growth { get; set; }
 }
