+++ LeagueOptimizer.Abstractions/Stats/IStat.cs
@@ -5,16 +5,16 @@ public interface IStat
     /// <summary>
     /// The base value of the stat
     /// </summary>
-    public double Base { get; set; }
+    public decimal Base { get; set; }
 
     /// <summary>
     /// The bonus value of the stat
     /// </summary>
-    public double Bonus { get; set; }
+    public decimal Bonus { get; set; }
 
     // todo make function or property
     /// <summary>
     /// The total value of the stat
     /// </summary>
-    public double Total => Base + Bonus;
+    public decimal Total => Base + Bonus;
 }
