+++ LeagueOptimizer.Abstractions/Stats/PerLevelStat.cs
@@ -2,12 +2,12 @@ using LeagueOptimizer.Abstractions.Champions.Data;
 
 namespace LeagueOptimizer.Abstractions.Stats;
 
-public class PerLevelStat(double baseValue, double growth) : IPerLevelStat
+public class PerLevelStat(decimal baseValue, decimal growth) : IPerLevelStat
 {
-    public double Base { get; set; } = baseValue;
-    public double Growth { get; set; } = growth;
-    public double Bonus { get; set; } = 0;
-    public double Total { get; set; } = 0;
+    public decimal Base { get; set; } = baseValue;
+    public decimal Growth { get; set; } = growth;
+    public decimal Bonus { get; set; } = 0;
+    public decimal Total { get; set; } = 0;
 
     public PerLevelStat(StatData data) : this(data.Base, data.Growth){}
 }
